website/requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.5
django-heroku==0.3.1
django-versatileimagefield==2.2
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.0.0
psycopg2==2.9.3
python-magic==0.4.24
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
